src/api/requirements.txt
fastapi
uvicorn[standard]
redis
python-dotenv
psycopg2-binary
